+++ docs/sdk/code-reference/pipelines/sources/spark/weather/weather_forecast_ecmwf.md
@@ -1,0 +1,1 @@
+::: src.sdk.python.rtdip_sdk.pipelines.sources.spark.weather.weather_forecast_ecmwf
