+++ docs/sdk/code-reference/pipelines/transformers/spark/weather/ECMWF/nc_ExtractGrid_to_weather_data_model.md
@@ -1,0 +1,1 @@
+::: src.sdk.python.rtdip_sdk.pipelines.transformers.spark.weather.ECMWF.nc_ExtractGrid_to_weather_data_model
