+++ mkdocs.yml
@@ -170,7 +170,9 @@ nav:
                       - Weather:
                           - Base Weather: sdk/code-reference/pipelines/sources/spark/weather/base_weather.md
                           - Weather Forecast API V1: sdk/code-reference/pipelines/sources/spark/weather/weather_forecast_api_v1.md
-                          - Weather Forecast API V1 Multi: sdk/code-reference/pipelines/sources/spark/weather/weather_forecast_api_v1_multi.md
+                          - Weather Forecast ECMWF V1 Multi: sdk/code-reference/pipelines/sources/spark/weather/weather_forecast_api_v1_multi.md
+                          - Base MARS ECMWF: sdk/code-reference/pipelines/sources/spark/weather/base_weather_ecmwf.md
+                          - Weather Forecast ECMWF: sdk/code-reference/pipelines/sources/spark/weather/weather_forecast_ecmwf.md
                   - Python:
                       - Delta: sdk/code-reference/pipelines/sources/python/delta.md
                       - Delta Sharing: sdk/code-reference/pipelines/sources/python/delta_sharing.md
@@ -190,6 +192,9 @@ nav:
                           - PJM To Meters Data Model: sdk/code-reference/pipelines/transformers/spark/iso/pjm_to_mdm.md
                       - Weather:
                           - Raw Forecast To Weather Data Model: sdk/code-reference/pipelines/transformers/spark/raw_forecast_to_weather_data_model.md
+                          - nc Forecast Extract Base To Weather Data Model: sdk/code-reference/pipelines/transformers/spark/weather/ECMWF/nc_ExtractBase_to_weather_data_model.md
+                          - nc Forecast Extract Point To Weather Data Model: sdk/code-reference/pipelines/transformers/spark/weather/ECMWF/nc_ExtractPoint_to_weather_data_model.md
+                          - nc Forecast Extract Grid To Weather Data Model: sdk/code-reference/pipelines/transformers/spark/weather/ECMWF/nc_ExtractGrid_to_weather_data_model.md
 
               - Destinations:
                   - Spark:
